fastapi_backend/pyproject.toml
[project]
name = "awesome-docify"
version = "1.0.0"
description = "Awesome Docify - AI assistant for painless documentation updates"
authors = [{ name = "Anish Navalgund", email = "anishk.navalgund@gmail.com" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "fastapi[standard]>=0.115.0,<0.116",
    "pydantic-settings>=2.5.2,<3",
    "sqlalchemy>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "watchdog>=5.0.0,<6.0.0",
    "openai>=1.98.0",
    "qdrant-client>=1.15.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-qdrant>=0.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=3.4.0,<4",
    "ruff>=0.1.0,<0.2",
    "watchdog>=5.0.3,<6",
    "python-dotenv>=1.0.1,<2",
    "pytest>=8.3.3,<9",
    "mypy>=1.13.0,<2",
    "pytest-cov>=6.2.1",
]

[tool.uv]
package = false

[tool.mypy]
python_version = "3.12"
files = ["app/**/*.py"]
follow_imports = "skip"
ignore_missing_imports = true
strict = true

[tool.hatch.build.targets.sdist]
include = ["commands"]

[tool.hatch.build.targets.wheel]
include = ["commands"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
